requirements.txt
streamlit
opencv-python
faiss-cpu
torch
torchvision
torchaudio
ultralytics
insightface
onnxruntime
Pillow
